+++ src/Message/Formatter.php
@@ -250,7 +250,7 @@ final class Formatter
     /**
      * Converts a value to a string.
      *
-     * @param mixed $value The value to convert
+     * @param mixed $value The value to convert.
      *
      * @return string Converted string.
      */